requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile
#
aiohttp==3.8.4
    # via -r requirements.in
aiosignal==1.3.1
    # via aiohttp
asgiref==3.7.2
    # via django
async-timeout==4.0.2
    # via aiohttp
attrs==23.1.0
    # via aiohttp
certifi==2023.5.7
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==3.1.0
    # via
    #   aiohttp
    #   requests
crispy-bootstrap5==0.7
    # via -r requirements.in
cryptography==41.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.2.1
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
    #   django-allauth
    #   django-crispy-forms
    #   django-database-url
django-allauth==0.54.0
    # via -r requirements.in
django-crispy-forms==2.0
    # via
    #   -r requirements.in
    #   crispy-bootstrap5
django-database-url==1.0.3
    # via -r requirements.in
frozenlist==1.3.3
    # via
    #   aiohttp
    #   aiosignal
idna==3.4
    # via
    #   requests
    #   yarl
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
oauthlib==3.2.2
    # via requests-oauthlib
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.7.0
    # via django-allauth
python-decouple==3.8
    # via -r requirements.in
python3-openid==3.2.0
    # via django-allauth
requests==2.31.0
    # via
    #   django-allauth
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sqlparse==0.4.4
    # via django
typing-extensions==4.6.3
    # via asgiref
urllib3==2.0.2
    # via requests
yarl==1.9.2
    # via aiohttp
